LR4/requirements.txt
absl-py==2.1.0
aiofile==3.9.0
aiofiles==24.1.0
aiogram==3.15.0
aiohttp==3.9.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
blobfile==3.0.0
caio==0.9.17
certifi==2024.12.14
charset-normalizer==3.4.0
chex==0.1.88
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
etils==1.11.0
filelock==3.16.1
fire==0.7.0
flax==0.10.2
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.0
humanize==4.11.0
idna==3.10
importlib_resources==6.5.2
jax==0.4.38
jaxlib==0.4.38
Jinja2==3.1.5
kiwisolver==1.4.8
llama_models==0.0.63
llama_stack==0.0.63
llama_stack_client==0.0.63
lxml==5.3.0
magic-filter==1.0.12
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml_dtypes==0.5.1
msgpack==1.1.0
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.2.1
opencv-python==4.10.0.84
opt_einsum==3.4.0
optax==0.2.4
orbax-checkpoint==0.11.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.3
pyaml==24.12.1
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.2
scipy==1.15.0
setuptools==75.6.0
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
tensorstore==0.1.71
termcolor==2.5.0
tesseract==0.1.3
tiktoken==0.8.0
tokenizers==0.21.0
toolz==1.0.0
tqdm==4.67.1
transformers==4.48.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
yarl==1.18.3
zipp==3.21.0
